# Pool *[]byte instead of []byte