@abhinav 2022-11-10
Superseded by https://github.com/go-git/go-git/commit/9490da0f86a12269abb2099e2ead1f20eec166d2